+++ src/components/ComicNavigation.tsx
@@ -1,5 +1,6 @@
 import { ConnectedProps, connect } from 'react-redux'
 
+import useLockedItem from '@hooks/useLockedItem'
 import {
     nextComicSelector,
     previousComicSelector,
@@ -10,6 +11,7 @@ import { AppDispatch, RootState } from '@store/store'
 const mapState = (state: RootState) => {
     return {
         settings: state.settings.values,
+        lockedToItem: state.comic.lockedToItem,
         currentComic: state.comic.current,
         latestComic: state.comic.latest,
         randomComic: state.comic.random,
@@ -20,8 +22,8 @@ const mapState = (state: RootState) => {
 
 const mapDispatch = (dispatch: AppDispatch) => {
     return {
-        setCurrentComic: (comic: number) => {
-            dispatch(setCurrentComic(comic))
+        setCurrentComic: (comic: number, locked: boolean) => {
+            dispatch(setCurrentComic(comic, { locked }))
         },
     }
 }
@@ -31,6 +33,8 @@ type PropsFromRedux = ConnectedProps<typeof connector>
 type ComicNavigationProps = PropsFromRedux & {}
 
 function ComicNavigation({
+    settings,
+    lockedToItem,
     currentComic,
     latestComic,
     previousComic,
@@ -38,14 +42,62 @@ function ComicNavigation({
     randomComic,
     setCurrentComic,
 }: ComicNavigationProps) {
+    const {
+        hasLockedItem,
+        lockedItem,
+        randomLockedItemComic,
+        refreshRandomLockedItemComic,
+    } = useLockedItem(currentComic, settings, lockedToItem)
+
+    let navigationData: {
+        first: number
+        firstTitle: string
+        previous: number
+        previousTitle: string
+        next: number
+        nextTitle: string
+        last: number
+        lastTitle: string
+        random: number
+        randomTitle: string
+    }
+    if (hasLockedItem) {
+        navigationData = {
+            first: lockedItem.first ?? currentComic,
+            firstTitle: `Go to first strip with ${lockedItem.shortName}`,
+            previous: lockedItem.previous ?? currentComic,
+            previousTitle: `Go to previous strip with ${lockedItem.shortName}`,
+            next: lockedItem.next ?? currentComic,
+            nextTitle: `Go to next strip with ${lockedItem.shortName}`,
+            last: lockedItem.last ?? currentComic,
+            lastTitle: `Go to last strip with ${lockedItem.shortName}`,
+            random: randomLockedItemComic ?? 0,
+            randomTitle: `Go to random strip with ${lockedItem.shortName}`,
+        }
+    } else {
+        navigationData = {
+            first: 1,
+            firstTitle: `Go to first strip`,
+            previous: previousComic,
+            previousTitle: `Go to previous strip`,
+            next: nextComic,
+            nextTitle: `Go to next strip`,
+            last: latestComic,
+            lastTitle: `Go to last strip`,
+            random: randomComic,
+            randomTitle: `Go to random strip`,
+        }
+    }
+
     return (
         <ul className="menu qc-ext qc-ext-navigation-menu" id="comicnav">
             <li>
                 <a
-                    href="view.php?comic=1"
+                    href={`view.php?comic=${navigationData.first}`}
+                    title={navigationData.firstTitle}
                     onClick={(e) => {
                         e.preventDefault()
-                        setCurrentComic(1)
+                        setCurrentComic(navigationData.first, hasLockedItem)
                     }}
                 >
                     First
@@ -53,10 +105,11 @@ function ComicNavigation({
             </li>
             <li>
                 <a
-                    href={'view.php?comic=' + previousComic}
+                    href={`view.php?comic=${navigationData.previous}`}
+                    title={navigationData.previousTitle}
                     onClick={(e) => {
                         e.preventDefault()
-                        setCurrentComic(previousComic)
+                        setCurrentComic(navigationData.previous, hasLockedItem)
                     }}
                 >
                     Previous
@@ -64,10 +117,11 @@ function ComicNavigation({
             </li>
             <li>
                 <a
-                    href={'view.php?comic=' + nextComic}
+                    href={`view.php?comic=${navigationData.next}`}
+                    title={navigationData.nextTitle}
                     onClick={(e) => {
                         e.preventDefault()
-                        setCurrentComic(nextComic)
+                        setCurrentComic(navigationData.next, hasLockedItem)
                     }}
                 >
                     Next
@@ -75,10 +129,11 @@ function ComicNavigation({
             </li>
             <li>
                 <a
-                    href={'view.php?comic=' + latestComic}
+                    href={`view.php?comic=${navigationData.last}`}
+                    title={navigationData.lastTitle}
                     onClick={(e) => {
                         e.preventDefault()
-                        setCurrentComic(latestComic)
+                        setCurrentComic(latestComic, hasLockedItem)
                     }}
                 >
                     {currentComic === latestComic ? 'Last' : 'Latest'}
@@ -86,10 +141,12 @@ function ComicNavigation({
             </li>
             <li>
                 <a
-                    href={'view.php?comic=' + randomComic}
+                    href={`view.php?comic=${navigationData.random}`}
+                    title={navigationData.randomTitle}
                     onClick={(e) => {
                         e.preventDefault()
-                        setCurrentComic(randomComic)
+                        setCurrentComic(navigationData.random, hasLockedItem)
+                        refreshRandomLockedItemComic()
                     }}
                 >
                     Random