+++ src/components/NavElement/NavElement.tsx
@@ -1,10 +1,12 @@
 import { ConnectedProps, connect } from 'react-redux'
 
 import NavButton from '@components/ComicDetailsPanel/NavButton/NavButton'
+import { ComicId } from '@models/ComicId'
 import { HydratedItemNavigationData } from '@models/HydratedItemData'
 import { ItemId } from '@models/ItemId'
 import { skipToken } from '@reduxjs/toolkit/dist/query'
 import { useRandomComicQuery } from '@store/api/itemApiSlice'
+import { setLockedToItem } from '@store/comicSlice'
 import { AppDispatch, RootState } from '@store/store'
 
 import { createTintOrShade } from '~/color'
@@ -15,18 +17,23 @@ const mapState = (state: RootState) => {
     return {
         settings: state.settings.values,
         currentComic: state.comic.current,
+        lockedToItem: state.comic.lockedToItem,
     }
 }
 
-const mapDispatch = (_dispatch: AppDispatch) => {
-    return {}
+const mapDispatch = (dispatch: AppDispatch) => {
+    return {
+        setLockedToItem: (item: ItemId | null) => {
+            dispatch(setLockedToItem(item))
+        },
+    }
 }
 
 const connector = connect(mapState, mapDispatch)
 type PropsFromRedux = ConnectedProps<typeof connector>
 type NavElementProps = PropsFromRedux & {
     item: HydratedItemNavigationData
-    onSetCurrentComic: (_: number) => void
+    onSetCurrentComic: (comic: ComicId, locked: boolean) => void
     useColors: boolean
     onShowInfoFor: (_: ItemId) => void
     mode: NavElementMode
@@ -45,6 +52,8 @@ export enum NavElementMode {
 export function NavElement({
     settings,
     currentComic,
+    lockedToItem,
+    setLockedToItem,
     item,
     useColors,
     onSetCurrentComic,
@@ -119,6 +128,8 @@ export function NavElement({
 
     const [alternateLayout, selfRef] = useAlternateLayout()
 
+    const locked = lockedToItem !== null && lockedToItem === item.id
+
     return (
         <>
             <div
@@ -144,14 +155,40 @@ export function NavElement({
                     comicNo={item.first !== currentComic ? item.first : null}
                     title={`First strip with ${item.shortName}`}
                     faClass="fast-backward"
-                    onSetCurrentComic={onSetCurrentComic}
+                    onSetCurrentComic={(c) => onSetCurrentComic(c, locked)}
                 />
                 <NavButton
                     comicNo={item.previous}
                     title={`Previous strip with ${item.shortName}`}
                     faClass="backward"
-                    onSetCurrentComic={onSetCurrentComic}
+                    onSetCurrentComic={(c) => onSetCurrentComic(c, locked)}
                 />
+                {settings?.showItemChainButton &&
+                    (locked ? (
+                        <button
+                            title={`Unlock page navigation from ${item.shortName}`}
+                            className={'flex-none px-2 block'}
+                            onClick={(e) => {
+                                e.preventDefault()
+                                setLockedToItem(null)
+                            }}
+                        >
+                            <i className={`fa fa-chain`}></i>
+                        </button>
+                    ) : (
+                        mode === NavElementMode.Present && (
+                            <button
+                                title={`Lock page navigation to ${item.shortName}`}
+                                className={'flex-none px-2 block'}
+                                onClick={(e) => {
+                                    e.preventDefault()
+                                    setLockedToItem(item.id)
+                                }}
+                            >
+                                <i className={`fa fa-chain-broken`}></i>
+                            </button>
+                        )
+                    ))}
                 <button
                     className={
                         'font-bold flex-auto py-1' +
@@ -174,13 +211,13 @@ export function NavElement({
                     comicNo={item.next}
                     title={`Next strip with ${item.shortName}`}
                     faClass="forward"
-                    onSetCurrentComic={onSetCurrentComic}
+                    onSetCurrentComic={(c) => onSetCurrentComic(c, locked)}
                 />
                 <NavButton
                     comicNo={item.last !== currentComic ? item.last : null}
                     title={`Last strip with ${item.shortName}`}
                     faClass="fast-forward"
-                    onSetCurrentComic={onSetCurrentComic}
+                    onSetCurrentComic={(c) => onSetCurrentComic(c, locked)}
                 />
                 {mode === NavElementMode.Present &&
                     settings?.showItemRandomButton && (
@@ -189,7 +226,7 @@ export function NavElement({
                             title={`Random strip with ${item.shortName}`}
                             faClass="question"
                             onSetCurrentComic={(c) => {
-                                onSetCurrentComic(c)
+                                onSetCurrentComic(c, locked)
                                 refreshRandomComic()
                             }}
                         />